backend/requirements.txt
asgiref>=3.6.0
django>=3.0
django-cors-headers>=3.7
django-storages>=1.11
python-dotenv>=0.15.0
requests>=2.25.0
anthropic>=0.1.0
openai>=0.27.0
boto3>=1.20.0
pymysql>=1.0.2
mysqlclient>=2.1.0
spacy[transformers]>=3.1.0
fastapi>=0.75.0
pydantic>=1.8.0